ML_VERSE_BACKEND/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
bcrypt==4.1.2
dnspython==2.4.2
Flask-JWT-Extended==4.4.4